authserver/src/main/java/com/example/authserver/controller/RegisterController.java
package com.example.authserver.controller;

import com.example.authserver.dto.RegisterUserDto;
import com.example.authserver.service.RegistrationService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class RegisterController {
    private final RegistrationService registrationService;

    public RegisterController(RegistrationService registrationService) {
        this.registrationService = registrationService;
    }

    @PostMapping("/register")
    public ResponseEntity<String> registerUser(@RequestBody RegisterUserDto registerUserDto) {
        return registrationService.registerUser(registerUserDto);
    }
}
